pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "xagent"
version = "0.1.0"
description = "Framework-agnostic enterprise AI agent skeleton"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "litellm>=1.65.8,<2.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "PyYAML>=6.0.2,<7.0.0",
]

[project.scripts]
xagent = "xagent.main:main"

[dependency-groups]
dev = [
  "pytest>=8.3.5,<9.0.0",
  "ruff>=0.11.0,<0.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/xagent"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.uv]
package = true
